Render the faceted parameters query box via Ajax.
This allows us to get the calendar start and end dates from fullcalendar.js and
means that the query box gets automatically updated when the calendar view is
changed.

This relies on a custom event that for the moment is triggered from within a
modified fullcalendar.js. Will have to find a permanent solution (either
override/monkey-patch or submit patch to fullcalendar.js author).